requirements.txt
requests==2.26.0
setuptools
python-dotenv
django==5.1.3
asgiref==3.8.1
sqlparse==0.5.2
spacy
spacy[lookups] @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
flair